feat!: Move progress event data to POJO object
Fixes #6.
